front end will dynamically render message objects from websocket
