+++ src/containers/FrontTester.tsx
@@ -2,40 +2,20 @@ import FrontLog from '../components/FrontLog';
 import { useState, useEffect } from 'react';
 
 
-const socket = new WebSocket('ws://localhost:1234');
 
 
 export default function FrontTester() {
-  const [message, updateMessage] = useState([]);
-
-
-  socket.addEventListener('open', (event) => {
-    console.log('CONNECTED TO WEB SOCKET FROM CLIENT Side');
-  });
-  
-  socket.addEventListener('message', (event) => {
-    // logic to display received data here
-    // likely use state components
-    console.log('MESSAGE RECEIVED FROM 1234: ', event.data);
-    updateMessage([...message, event.data]);
-  });
-  
-  const sendMessage = () => {
-    socket.send('1. CLIENT 1 JUST SEND THIS MESSAGE TO SERVER!!!!');
-  };
-
-  useEffect(() => {}, []);
   
   return (
     <div className='bg-gray-900 h-screen'>
       <FrontLog />
-      <button onClick={sendMessage} style={{ color: 'white' }}>
+      {/* <button onClick={sendMessage} style={{ color: 'white' }}>
                 SEND MESSAGE
       </button>
-      {message.map((element) => {
+      {message.map((element, index) => {
         return (
         <div style={{color: 'white'}}>{element}</div>
-        )})}
+        )})} */}
     </div>
     // <div className="front-tester-container">
     //   <div className="api-dropdown"></div>
